refactor: separate orchestrators from agents for clearer architecture
Move orchestration logic out of agents/ into dedicated orchestrators/ folder
to clearly distinguish between agents (autonomous LLM-powered entities) and
orchestrators (coordination logic).

Changes:
- Create loan_avengers/orchestrators/ folder
- Move conversation_orchestrator.py to orchestrators/
- Move conversation_state_machine.py to orchestrators/
- Move sequential_pipeline.py to orchestrators/
- Update all imports in app.py and orchestrators
- Add orchestrators/__init__.py with clean exports

Folder structure now:
- loan_avengers/agents/ - Pure agents only (intake, credit, income, risk)
- loan_avengers/orchestrators/ - Coordination layer (conversation, pipeline)

Rationale:
- Orchestrators coordinate agents but are NOT agents themselves
- ConversationOrchestrator bridges UI → LoanApplication → Pipeline
- ConversationStateMachine manages deterministic conversation flow
- SequentialPipeline orchestrates multi-agent workflows

All tests passing (13/13) after reorganization.
